Added the Book Model and implemented the methods on BooksController
Created a Models Book with 4 parameters (id, name, price and author) and a "add", "remove", "findById" and "allBooks" methods. Implemented the index method (which calls the view index), the create method (which opens the view create), the save method(which adds a Book from a Form we created, and redirects to index view), the edit method(which calls the edit view), the update method(which changes a book receiving its data from a Form and redirects to index), the destroy method(which eliminates a book and redirects to index) and a show method(which opens the show view) on BooksController.