Added views
Added to the views package the index view, and the layout view(which is used in the views inserted into the books package.
Added to the books package a create view(which receives data and calls the method save in BooksController), a edit view(which receives data and calls the method edit in BooksController), a index view(which shows all the books and opens a button to call the create method in BooksController and a href to call the method show in BooksController. and a show view(which shows a particular book and opens 2 buttons, one to call the method edit in BooksController and the other to call the method destroy in BooksController